requirements.txt
# 基础依赖
numpy>=1.24.0
Pillow>=9.5.0